refactor: move detail-modal flows to DetailController (break-up phase 2)
Extract the job/node/user/account detail-modal logic off SlurmMonitor
into stoei/detail_controller.py. DetailController talks to the app
through a minimal _DetailHost Protocol (no circular import, no Any, no
TYPE_CHECKING). SlurmMonitor keeps thin delegators so call sites, action
bindings, and tests are unchanged.

app.py: 2771 -> 2423 lines.
